getEvents: batch getBlockNumber and getLogs in one request
Patch `getEvents` to use JSON-RPC 2.0 batch requests to call getBlockNumber('latest') and getLogs(fromBlock, toBlock) in one
request-response cycle. This ensures that the block number and logs are from the same RPC node.
Since `eth_getLogs` does not return an error if `toBlock` is larger than largest block number known to the RPC node,
  getBlockNumber('latest') being served by a node synced to block 100, then
  getLogs(fromBlock: 99, toBlock: 100) being served by a node synced to block 99
could make the caller miss the logs contained in block 100.
The caller would think it has processed the logs in block 100 when it really has not